[FIX] fix issues with clippy errors and rename files
