+++ parquet-variant/examples/variant_to_json_examples.rs
